chore: refactor: restructure file structure, to clarify where code is actually run (parent vs child process)
