+++ src/parent-process/README.md
@@ -1,0 +1,1 @@
+The code in this folder is executed in the parent process
