+++ src/shared/README.md
@@ -1,0 +1,1 @@
+The code in this folder is executed BOTH in the parent process and child processes
