+++ src/shared/lib.ts
